Enable compatibility with bukkit plugins
By swapping out the spigot-api with bukkit, and changing the method InventoryClickEvent#getClickedInventory to the standard InventoryEvent#getInventory, we can enable compatibility with bukkit plugins while preserving the same functionality. This is because the #getInventory method simply returns the top level inventory, which is all that is ever of interest for this library.